more warnings cleanup

+++ dad-is-great-in-C/lastword.cpp
@@ -9,7 +9,7 @@ int main() {
   getline(cin, line);
   size_t end = line.find_last_not_of(" ");
   string trimmed = line.substr(0, end + 1);
-  string last(trimmed.begin() + trimmed.rfind(" ") + 1, trimmed.end());
+  string last(trimmed.begin() + static_cast<int64_t>(trimmed.rfind(" ")) + 1, trimmed.end());
   cout << last;
   return 0;
 }
